Extract pack.rs into pack/{mod,render,impact}.rs submodules
- pack/mod.rs: core types (Preset, Format, Section, Pack, PackBuilder),
  SectionWants, build() orchestration, collector helper functions, tests
- pack/render.rs: standalone render functions + utilities
  (render_diff_ordered, render_diagnostics, render_tests, render_entry,
   render_map, lang_for_path, mk_section, project_name)
- pack/impact.rs: impact-specific rendering
  (try_collect_scoped, try_collect_scoped_findings,
   try_collect_per_finding, format_file_header)

Reduces pack/mod.rs from 1437 to ~550 lines. Render and impact
logic is now self-contained in their own modules.

Closes #7
